* Added governace env check
