+++ lib/cli/scripts/plugins/plugins-model.ts
